requirements.txt
# AdsenseAI Campaign Risk Analyzer - Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Text Analysis
textblob==0.17.1
vaderSentiment==3.3.2

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Google Generative AI (Gemini)
google-generativeai==0.3.2

# Environment Variables
python-dotenv==1.0.0

# CORS Support
python-jose[cryptography]==3.3.0
